+++ fe/src/page/plan/PlanAdd.jsx
@@ -5,7 +5,7 @@ import { useNavigate } from "react-router-dom";
 import { toaster } from "../../components/ui/toaster.jsx";
 import { Modal } from "/src/components/root/Modal.jsx";
 import { Breadcrumb } from "../../components/root/Breadcrumb.jsx";
-import { GoogleMapsContainer } from "./GoogleMapsContainer.jsx";
+import GoogleMapsContainer from "./GoogleMapsContainer.jsx";
 
 function PlanAdd(props) {
   const [backToListModalOpen, setBackToListModalOpen] = useState(false);
